ccTimer updated
Relies on QTime only now (and no more on QElapsedTimer which is quite
buggy... or at least not what expected!)

+++ libs/qCC_db/ccTimer.h
@@ -49,9 +49,9 @@ public:
     static void Init();
 
     //! Returns number of seconds since timer was initialized
-    static qint64 Sec();
+    static int Sec();
     //! Returns number of milli-seconds since timer was initialized
-    static qint64 Msec();
+    static int Msec();
 
 protected:
